Consolidate modules - media to it's own folder - various files into utils folder - graphql-request-client into graphql folder - rename testData -> test-data
