+++ packages/sitecore-jss/src/graphql/graphql.test.ts
@@ -1,10 +1,10 @@
 /* eslint-disable no-unused-expressions */
 import { expect } from 'chai';
 import { getAppRootId, siteNameError, languageError } from './app-root-query';
-import { GraphQLRequestClient } from '../graphql-request-client';
-import appRootQueryResponse from '../testData/mockAppRootQueryResponse.json';
+import { GraphQLRequestClient } from './graphql-request-client';
+import appRootQueryResponse from '../test-data/mockAppRootQueryResponse.json';
 import nock from 'nock';
-import { SitecoreTemplateId } from '../constants';
+import { SitecoreTemplateId } from '../utils/constants';
 
 // Todo: Instead of creating a GraphQLRequestClient instance, we can use the
 // GraphQLClient interface. This would be a better mock. Potentially create a
