+++ packages/sitecore-jss/src/graphql/index.ts
@@ -1,5 +1,10 @@
 export { getAppRootId, AppRootQueryResult } from './app-root-query';
 export {
+  GraphQLClient,
+  GraphQLRequestClient,
+  GraphQLRequestClientConfig,
+} from './graphql-request-client';
+export {
   SearchQueryResult,
   SearchQueryVariables,
   SearchServiceConfig,
