+++ packages/sitecore-jss/src/index.ts
@@ -1,19 +1,6 @@
 // NOTE: all imports are now named as to not make breaking changes
 // and to keep react-native working with cjs modules.
-// This will very likely change again when sub-modules are added.
 
-import * as mediaApi from './media-api';
-import * as constants from './constants';
-
-export { mediaApi, constants };
-export { default as debug, Debugger } from './debug';
-export { HttpDataFetcher, HttpResponse, fetchData } from './data-fetcher';
-export {
-  GraphQLClient,
-  GraphQLRequestClient,
-  GraphQLRequestClientConfig,
-} from './graphql-request-client';
-export { AxiosDataFetcher, AxiosDataFetcherConfig } from './axios-fetcher';
 export {
   AppRootQueryResult,
   SearchQueryResult,
@@ -21,7 +8,12 @@ export {
   SearchQueryVariables,
   SearchServiceConfig,
   getAppRootId,
+  // grapql-request-client
+  GraphQLClient,
+  GraphQLRequestClient,
+  GraphQLRequestClientConfig,
 } from './graphql';
+
 export {
   ExperienceEditor,
   HorizonEditor,
@@ -31,9 +23,27 @@ export {
   resetEditorChromes,
   resetExperienceEditorChromes,
   resolveUrl,
+  // debug
+  debug,
+  Debugger,
+  // axios-fetcher
+  AxiosDataFetcher,
+  AxiosDataFetcherConfig,
+  // cache-client
+  CacheClient,
+  CacheOptions,
+  MemoryCacheClient,
+  // data-fetcher
+  HttpDataFetcher,
+  HttpResponse,
+  fetchData,
+  // constants
+  constants,
 } from './utils';
-export { CacheClient, CacheOptions, MemoryCacheClient } from './cache-client';
 
+export { mediaApi } from './media';
+
+// dictionary-service (i18n)
 export {
   DictionaryPhrases,
   DictionaryService,
