+++ packages/sitecore-jss/src/layout/graphql-layout-service.ts
@@ -1,7 +1,7 @@
 import { LayoutServiceBase } from './layout-service';
 import { LayoutServiceData } from './models';
-import { GraphQLClient, GraphQLRequestClient } from '../graphql-request-client';
-import debug from '../debug';
+import { GraphQLClient, GraphQLRequestClient } from '../graphql/graphql-request-client';
+import debug from '../utils/debug';
 
 export type GraphQLLayoutServiceConfig = {
   /**
