+++ packages/sitecore-jss/src/media/index.ts
@@ -1,0 +1,3 @@
+﻿import * as mediaApi from './media-api';
+
+export { mediaApi };
