+++ packages/sitecore-jss/src/media/media-api.ts
