+++ packages/sitecore-jss/src/test-data/mockAppRootQueryResponse.json
