+++ packages/sitecore-jss/src/test-data/mockDictionaryQueryResponse.json
