+++ packages/sitecore-jss/src/utils/constants.ts
