+++ packages/sitecore-jss/src/utils/index.ts
@@ -1,6 +1,17 @@
 import isServer from './is-server';
 import resolveUrl from './resolve-url';
+import * as constants from './constants';
+import { default as debug, Debugger } from './debug';
+import { AxiosDataFetcher, AxiosDataFetcherConfig } from './axios-fetcher';
+import { HttpDataFetcher, HttpResponse, fetchData } from './data-fetcher';
+import { CacheClient, CacheOptions, MemoryCacheClient } from './cache-client';
+
+export { constants };
+export { debug, Debugger };
 export { isServer, resolveUrl };
+export { AxiosDataFetcher, AxiosDataFetcherConfig };
+export { HttpDataFetcher, HttpResponse, fetchData };
+export { CacheClient, CacheOptions, MemoryCacheClient };
 export {
   ExperienceEditor,
   HorizonEditor,
